Update library versions in gradle configuration
The version numbers of several libraries in the Gradle configuration have been updated. Specifically, the Compose, kotlinx-serialization, coroutines and ktor libraries have been updated to their most recent stable or release candidate versions. This update ensures the project's dependencies are up-to-date and helps maintain compatibility with other software and libraries.