Rename RegisterUser to RegisterBody and add login functionality
The class "RegisterUser" was renamed to "RegisterBody". This has led to changes in various areas of the code notably in the "authentication" route, the "AuthenticationRepository", and related services and validators. Additionally, a new functionality of user login was introduced with its respective body defined as "LoginBody". This involves making data as received in the request for authentication and generating a response accordingly.