Add user authentication to ZoneRepository methods
Modified the methods in ZoneRepository to require a User input, thus ensuring that only authorized users can manipulate the zone operations. This update augments the security of the application as it confirms the user has the right to perform the requested operations (fetch, update, delete, addRRSet, updateRRSet, deleteRRSet) on a given zone.