Remove main.kt and refactor UI package files
This commit removes the main.kt file from the 'desktopMain' package. It also reorganizes and renames classes within the 'ui' package, moving them into a new 'theme' package. Additionally, a new class Res.kt is added into a newly created 'state' package. The purpose of these changes is to improve the structure and readability of the codebase.