Add registration screen and navigation
Updated the LoginScreen.kt to implement a navigational feature to RegisterScreen.kt. Added navigation to both screens allowing the user to switch between the login screen and the registration screen. Also, a RegisterScreen is created with similar functionality as LoginScreen.kt, providing a UI for new users to create an account.