Refactor strings resource and update dependencies
Moved strings resources from the "launchpad" module to the "commonCompose" module and updated Compose and AndroidX versions. Additionally, the "app" module was added to settings.gradle.kts and the new module's build file structure was introduced. This change provides a unified location for storing string resources across modules and keeps the dependencies up to date.