Update JVM toolchain, plugin version, and upgrade dependencies
JVM toolchain has been updated to use libs.versions.java and vendor was set to JvmVendorSpec.AZUL on shared, pdnsClient and server build files. Foojay resolver convention plugin was added to settings.gradle.kts. Java version in libs.versions file has been bumped up to 17 and ktor version has been updated to 2.3.10. This improves the toolchain control and keeps the project up-to-date with the latest stable versions of the plugins and dependencies.