Update security settings and switch to local Postgres
This commit applies crucial updates to our security settings in Security.kt and switches the database url to a localized one in application.conf. Additionally, the Java version was updated and the docker image used for the jib was specified under the Ktor section of the build script to keep consistency across project.